requirements.txt
absl-py==2.1.0
alembic==1.14.0
amqp==5.3.1
aniso8601==9.0.1
appnope==0.1.3
asgiref==3.7.2
asttokens==2.4.1
astunparse==1.6.3
attrs==23.2.0
beautifulsoup4==4.12.2
bidict==0.22.1
billiard==4.2.1
blinker==1.7.0
bs4==0.0.1
celery==5.4.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
clips==1.1.0
clipspy==1.0.3
colorlog==6.9.0
comm==0.2.0
contourpy==1.2.0
cron-descriptor==1.4.5
cryptography==43.0.3
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
deep-translator==1.11.4
Django==5.0.2
django-celery-beat==2.7.0
django-ckeditor==6.7.1
django-cors-headers==4.3.1
django-js-asset==2.2.0
django-timezone-field==7.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
et-xmlfile==1.1.0
executing==2.0.1
experta==1.9.4
filelock==3.13.3
Flask==3.0.0
Flask-Cors==4.0.0
Flask-Mail==0.9.1
Flask-RESTful==0.3.10
Flask-SocketIO==5.3.6
flatbuffers==24.3.25
fonttools==4.46.0
frozendict==1.2
fsspec==2024.3.1
gast==0.6.0
google-pasta==0.2.0
graphviz==0.20.1
grpcio==1.68.0
h11==0.14.0
h5py==3.12.1
huggingface-hub==0.22.2
idna==3.6
imbalanced-learn==0.12.2
imblearn==0.0
inflection==0.5.1
install==1.3.5
ipykernel==6.27.1
ipython==8.18.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jupyter_client==8.6.0
jupyter_core==5.5.0
keras==3.6.0
kiwisolver==1.4.5
kombu==5.4.2
libclang==18.1.1
Mako==1.3.6
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.4.1
mlxtend==0.23.0
mpmath==1.3.0
mysqlclient==2.2.4
namex==0.0.8
nest-asyncio==1.5.8
networkx==3.3
numpy==1.26.2
opencv-python==4.9.0.80
openpyxl==3.1.2
opt_einsum==3.4.0
optree==0.13.1
optuna==4.1.0
outcome==1.3.0.post0
packaging==23.2
pandas==2.1.3
parso==0.8.3
pexpect==4.9.0
Pillow==10.1.0
platformdirs==4.1.0
prompt-toolkit==3.0.41
protobuf==5.28.3
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.22
Pygments==2.17.2
PyJWT==2.8.0
PyMySQL==1.1.0
pyodbc==5.1.0
pyparsing==3.1.1
pypi==2.1
PySocks==1.7.1
python-crontab==3.2.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.8.0
python-socketio==5.10.0
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
redis==5.2.1
regex==2023.12.25
requests==2.31.0
rich==13.9.4
safetensors==0.4.2
schema==0.6.7
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.0
selenium==4.21.0
simple-websocket==1.0.0
six==1.16.0
smogn==0.1.2
sniffio==1.3.1
snownlp==0.12.3
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.36
sqlparse==0.4.4
stack-data==0.6.3
sympy==1.12
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
threadpoolctl==3.2.0
tokenizers==0.15.2
torch==2.2.2
tornado==6.4
tqdm==4.66.2
traitlets==5.14.0
transformers==4.39.3
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.11.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.12
webdriver-manager==4.0.1
Werkzeug==3.0.1
wrapt==1.16.0
wsproto==1.2.0
xgboost==2.0.3
